Much bugfixing, almost all of them trivial.
This thing was a mess when I started, so there were a pile
of things sitting around that needed to be cleaned up.

Type errors, misspellings, lots of global state that needed
to be encapsulated or else parametrized in some other way...
special data structures for set representation and one of the
stranger hash tables (more a symbol table, really) that i've
yet seen.

Most of it is done now, when I get the energy to really go
through and un-fuck the code that prints the yacc driver,
plus the other printing routines, and tighten up the set
module, things should be almost ready to go.